<H2/>'s children that is a component blocks highlighting
